requirements.txt
autopep8==2.0.4
click==8.1.7
cv==1.0.0
joblib==1.3.2
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pycodestyle==2.11.0
pytesseract==0.3.10
regex==2023.10.3
tomli==2.0.1
tqdm==4.66.1
Wand==0.6.11
